Move serializable types into their own mod. (#46)
