+++ src/storage/mod.rs
@@ -3,4 +3,5 @@ pub mod db;
 pub mod lsm;
 pub mod serde;
 pub mod sstable;
+pub mod types;
 pub mod utils;
