+++ src/storage/types/mod.rs
@@ -1,0 +1,3 @@
+mod datum;
+
+pub use datum::*;
